+++ Options/CHANGELOG.lua
@@ -3,6 +3,10 @@ if not TMW then return end
 TMW.CHANGELOG_LASTVER="7.4.0"
 
 TMW.CHANGELOG = [==[
+## v10.0.2
+### Bug Fixes
+* Fix #1984 - attempt to index local 'conditionData' (a nil value)
+
 ## v10.0.1
 ### Bug Fixes
 * Fix #1974 - Assorted warnings about XML attributes